+++ lib/src/theme/src/dark_theme.dart
@@ -48,7 +48,6 @@ ThemeData generateDarkTheme({
   /// ```
   AppFont? bodyFont,
 }) {
-  bool isIOS = !kIsWeb && Platform.isIOS;
   MaterialColor color = getThemeColor(
     theme: theme,
     color: mainColor,
@@ -101,10 +100,7 @@ ThemeData generateDarkTheme({
       contentPadding: const EdgeInsets.all(10),
       filled: true,
       fillColor: Colors.grey.shade800,
-      border: UnderlineInputBorder(
-        borderRadius: BorderRadius.circular(10),
-        borderSide: BorderSide.none,
-      ),
+      border: const ThemedInputBorder(),
       labelStyle: TextStyle(
         color: Colors.grey.shade400,
       ),
@@ -139,17 +135,17 @@ ThemeData generateDarkTheme({
 
     // Header / AppBar
     secondaryHeaderColor: color,
-    appBarTheme: AppBarTheme(
+    appBarTheme: const AppBarTheme(
       systemOverlayStyle: SystemUiOverlayStyle(
         // Changes the status bar colors and buttons (Top)
-        statusBarColor: kDarkBackgroundColor,
-        statusBarIconBrightness: isIOS ? Brightness.dark : Brightness.light,
-        statusBarBrightness: isIOS ? Brightness.dark : Brightness.light,
+        statusBarColor: kDarkBackgroundColor, // Android Only
+        statusBarIconBrightness: Brightness.light, // Android Only
+        statusBarBrightness: Brightness.dark, // iOS Only
 
         // Changes the navigation bar colors and buttons
-        systemNavigationBarColor: kDarkBackgroundColor,
-        systemNavigationBarDividerColor: kDarkBackgroundColor,
-        systemNavigationBarIconBrightness: isIOS ? Brightness.dark : Brightness.light,
+        systemNavigationBarColor: kDarkBackgroundColor, // Android only
+        systemNavigationBarDividerColor: kDarkBackgroundColor, // Android only
+        systemNavigationBarIconBrightness: Brightness.light, // Android only
       ),
       color: kDarkBackgroundColor,
       surfaceTintColor: kDarkBackgroundColor,