+++ lib/src/theme/theme.dart
@@ -9,9 +9,11 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:layrz_models/layrz_models.dart';
 import 'package:layrz_theme/src/helpers/helpers.dart';
 import 'package:http/http.dart' as http;
+import 'dart:math' as math;
 
 part 'src/platform.dart';
 part 'src/light_theme.dart';
 part 'src/dark_theme.dart';
 part 'src/constants.dart';
 part 'src/font.dart';
+part 'src/utilities.dart';